+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [vNext]
+### Added
+- Allow to disable fetching page resources by dom snapshot script [Trello 2242](https://trello.com/c/C1YodBjt)
+
 ## [4.10.0] - 2020-10-16
 ### Added
 - Supporting check full element with ufg [Trello 2145](https://trello.com/c/8tPAnz66)